Login & Out Redirect URIs
Now they both return to localhost. Which is fine for now, as the website isn't being hosted anywhere.

Going to have to figure out which of the tokens are most condusive to storing in a Database to ensure that the user queries the right larder when navigating to the larder screen