Admit Defeat Vis-a-Vis csrf.
A short summary of the {% csrf token %} django tag: This gets translated into a csrftmiddleware input field, with a csrf token as the value of this input field. If you can grab input value fields, keeping track of this shouldn't be too hard: I did so via a use state. This csrf field then has to be passed back into django via the X_CSRFTOKEN field.

But Wait! Why does the django documentation ask for cookie finangling? [ https://docs.djangoproject.com/en/5.2/howto/csrf/ ] When itiallly GET-ing, the response contains a Set-Cookie header. In order for Django to accept a csrf-ed POST request, you require BOTH the input csrf middleware field AND the value from this newly-set cookie.
At least for my browser, Set-Cookie is blocked if the Samesite is "Lax" and the GET is not a navigation request (Which it is not, I'm GET-ing via fetch).
If I set Samesite to be "None", then Set-Cookie is blocked because localhost isn't https.
Either way, set-cookie is blocked, so I cannot send Django the token it needs to accept me as legitimate via CSRF.
Hence, I depose CSRF, before it Deposes me!