Get-ing & Posting now done by a Context Provider
My golly! look how much cleaner this is!
Furthermore, hopefully by posting the  post function in one place, it'll be easier to set up csrf in the indeterminable future.
It's a white whale. I am less fanatical then Ahab, so I will leave CSRF to rest for now.
My memory is long, however, and my lust for revenge is great.