Register Admin.
Good news! In the beginning, it worked! I could create a superuser, and gain admin access to my website. Neat!
Bad News! I could create A superuser. A single superuser. If I tried to create a second user, I get "error: the username is not unique"
Huh? I set the username field to the email, the e-mail should be unique. What's going on?