CSRF Works Now?
That's all I had to do for SameSite: None to work? Proclaim that it's secure?
I guess that makes sence, considering localhost runs on my local machine,. I wasted, what, hours trying to make an https localhost. And all I had to do was to set the Secure Settings to True?

In sum: How to do CSRF: Make sure Set-Cookie works.. In my case, because Localhost for some reason != localhost, I has to put my cookie's SameSite configuration to be None. In order for my browser not to block this SameSite: None, The cookie has to demand to be carried over a secure Network. Setting the COOKIE SECURE setting for Django sufficed; my connection was localhost, and localhost is special =)
When Posting:
1. Grab from the {% csrf_token %} input feild.
2. Grab cookie By using the code ever Django CSRF tutorial tells you to grab.
3. Make sure that  the X-CSRFToken Header contains the string you grabbed earlier taken from the {% csrf_token %} input feild.
4. Ensure the CSRF cookie is also sent. As most of my suffering entailed receiving the proper cookie, and not sending it, I haven't put much thought into this. I guess the "Credentials : include" is what's needed for that.