Stopped FoodItem Complaining about Unseen Yet Existing Foreign Key
I Kid you not, The food Item was searching for the no-longer-existing url_ext. It kicked up so much of a fuss that it utterly derailed the memo/migration on the name change.
This was fixed by setting the FoodItem's Foreign Key to an Integer Field, and then setting it back to a foreign key. And Altering the orderings of migrations.

Or, at least, that's what I hope. When running migrate, I first got a circular dependency error, and then a "=) everything is done already, dont't worry"