Larder: Not Yet a Context
Working First version of the Larder Front-end.

I'm looking at the "Total Refresh" (Wipe a Larder) portion of my plan.
Yeah, this is going to be hard without the larder having a designated Context: I'll have to alter the larder via the toolbar, a component only tangentially related to the FoodItemTable.